+++ R/tinify_defaults.R
@@ -7,8 +7,6 @@
 #'If called without any arguments, `tinify_defaults()` will print
 #'the current default options set to the console.
 #'
-#'To set any option back to the package default, set it to `NULL`.
-#'
 #'@param overwrite Boolean, defaults to `FALSE`. By default, tinify will create
 #'  a new file with the suffix '_tiny' and preserve the original file. Set
 #'  `TRUE` to instead overwrite the original file, with the same file name.
@@ -21,22 +19,23 @@
 #'  on file names, amount of file size reduction (% and Kb), and the number of
 #'  TinyPNG API calls made this month. If set to `TRUE`, tinify displays no
 #'  messages as it shrinks files.
-#'@param return_path String, optional. One of "`proj`", "`rel`", "`abs`", or
+#'@param return_path String or `NULL`, optional. One of "`proj`", "`rel`", "`abs`", or
 #'  "`all`". If "`proj`", will return the file path of the newly tinified image
 #'  file relative to the Rstudio project directory (looking for an .Rproj file).
 #'  If no project can be identified, returns `NA`. If "`rel`", will return the
 #'  file path of the newly tinified image file, relative to the \strong{current}
 #'  working directory at the time `tinify()` is called. If "`abs`", will return
 #'  the absolute file path of the newly tinified image file. If "`all`", will
-#'  return a named list with all file paths.
-#'@param resize Named list, optional. A named list with the elements `method` as
+#'  return a named list with all file paths. Set to `NULL` to stop returning a
+#'  file path.
+#'@param resize Named list or `NULL`, optional. A named list with the elements `method` as
 #'  a string, and `width` and/or `height` as numerics. Please note you can only
 #'  reduce an image's dimensions and make an image smaller with TinyPNG API, not
 #'  make an image larger. Method must be set to one of "scale", "fit", "cover",
 #'  or "thumb". If using "scale", you only need to provide `width` OR `height`,
 #'  not both. If using any other method, you must supply both a `width` AND
 #'  `height`. See <https://tinypng.com/developers/reference#resizing-images> and
-#'  the examples for more.
+#'  the examples for more. Set to `NULL` to stop resizing.
 #'@seealso [tinify()] to shrink image filesizes
 #'@seealso [tinify_key()] to set a default TinyPNG.com API key
 #'@export
@@ -44,10 +43,6 @@
 #' \dontrun{
 #'
 #' tinify_defaults(quiet = TRUE, suffix = "_small")
-#'
-#' # set an option back to the package default
-#'
-#' tinify_defaults(quiet = NULL)
 #'
 #' # show current defaults set
 #'
@@ -65,116 +60,121 @@ tinify_defaults <- function(overwrite,
                        return_path,
                        resize) {
 
-  # overwrite ==================================================================
-  if(!missing(overwrite)){
-    if(!identical(overwrite, TRUE) & !identical(overwrite, FALSE) & !identical(overwrite, NULL)) {
-      stop("Please provide 'overwrite' as TRUE or FALSE, or NULL to reset to package default")
-    } else {
-      options(tinify.overwrite = overwrite)
-    }
-    message(glue::glue("Tinify 'overwrite' changed to: {getOption('tinify.overwrite', default = FALSE)}"))
-  }
+  #  print defaults if all arguments missing ===================================
 
-  # suffix =====================================================================
-  if(!missing(suffix)){
-    if (is.null(suffix)) {
-      options(tinify.suffix = suffix)
-    } else if(!is.character(suffix) || length(suffix) > 1 || suffix == "") {
-      stop("Please provide 'suffix' as a non-empty character string, or NULL to reset to package default")
-    } else {
-      options(tinify.suffix = suffix)
-    }
-    suf_set <- if (!is.null(getOption('tinify.suffix'))) glue::glue("\"{getOption('tinify.suffix')}\"") else '\"_tiny\"'
-    message(glue::glue("Tinify 'suffix' changed to: {suf_set}"))
-  }
-
-  # quiet ======================================================================
-  if(!missing(quiet)){
-    if(!identical(quiet, TRUE) & !identical(quiet, FALSE) & !identical(quiet, NULL)) {
-      stop("Please provide 'quiet' as TRUE or FALSE, or NULL to reset to package default")
-    } else {
-      options(tinify.quiet = quiet)
-    }
-    message(glue::glue("Tinify 'quiet' changed to: {getOption('tinify.quiet', default = FALSE)}"))
-  }
-
-  # return_path ================================================================
-  if(!missing(return_path)){
-    if (is.null(return_path)) {
-      options(tinify.return_path = return_path)
-    } else if(length(return_path) > 1 || !(return_path %in% c("proj", "rel", "abs", "all"))) {
-      stop("Please provide 'return_path' as one of  'proj', 'rel', 'abs', or 'all', or NULL to reset to package default")
-    } else {
-      options(tinify.return_path = return_path)
-    }
-    ret_set <- if (!is.null(getOption('tinify.return_path'))) glue::glue("\"{getOption('tinify.return_path')}\"") else 'No return'
-    message(glue::glue("Tinify 'return_path' changed to: {ret_set}"))
-  }
-
-  # resize =====================================================================
-  if(!missing(resize)){
-
-    if (is.null(resize)) {
-      options(tinify.resize = resize)
-    } else if(!is.list(resize) | length(resize) < 2 | length(resize) > 3){
-      # Check resize is a list of min length 2 and max length 3
-      stop("Resize must be a list that includes a 'method' and one or both of 'width' or 'height'")
-    } else if(!("method" %in% names(resize) & ("width" %in% names(resize) | "height" %in% names(resize)))) {
-      # Check 'method' and at least one of 'width' or 'height' are named in resize
-      stop("Resize must be a list that includes a 'method' and one or both of 'width' or 'height'")
-    } else if(!(resize$method %in% c("fit", "scale", "cover", "thumb"))) {
-      # Check resize method is a string specifying one of the available options
-      stop('Method must be one of "fit", "scale", "cover" or "thumb"')
-    } else if("width" %in% names(resize) & !is.numeric(resize$width) | "height" %in% names(resize) & !is.numeric(resize$height)) {
-      # Check width and/or height are numbers
-      stop("Width and/or height must be a number")
-    } else if(identical(resize$method, "scale") & "width" %in% names(resize) & "height" %in% names(resize)) {
-      # Check only one of width or height provided for method 'scale'
-      stop("You must provide a width OR height for method 'scale', not both")
-    } else if(!identical(resize$method, "scale") & (!("width" %in% names(resize)) | !("height" %in% names(resize)))) {
-      # Check both width and height provided for other methods besides 'scale'
-      stop(paste0("You must provide a width AND height for method '", resize$method, "'"))
-    } else {
-      options(tinify.resize = resize)
-    }
-    if(!is.null(resize)) {
-      for (i in 1:length(resize)) {
-          if (is.character(resize[[i]])) {
-            message(glue::glue('Tinify \'resize\' {(names(resize)[i])} changed to: \"{resize[[i]]}\"'))
-          } else {
-            message(glue::glue('Tinify \'resize\' {(names(resize)[i])} changed to: {resize[[i]]}'))
-          }
-        }
-    } else {
-      message(glue::glue("Tinify 'resize' changed to: No resize"))
-    }
-  }
-
-  # print defaults =============================================================
   if (missing(overwrite) & missing(suffix) & missing(quiet) & missing(return_path) & missing(resize)) {
-    # overwrite
-    message(glue::glue("Tinify 'overwrite' set to: {getOption('tinify.overwrite', default = FALSE)}"))
-    # suffix
-    suf_set <- if (!is.null(getOption('tinify.suffix'))) glue::glue("\"{getOption('tinify.suffix')}\"") else '\"_tiny\"'
-    message(glue::glue("Tinify 'suffix' set to: {suf_set}"))
-    # quiet
-    message(glue::glue("Tinify 'quiet' set to: {getOption('tinify.quiet', default = FALSE)}"))
-    # return_path
-    ret_set <- if (!is.null(getOption('tinify.return_path'))) glue::glue("\"{getOption('tinify.return_path')}\"") else 'No return'
-    message(glue::glue("Tinify 'return_path' set to: {ret_set}"))
-    # resize
+
+    .tinify_print_defaults()
+
+  } else {
+
+    # Error checking ===========================================================
+
+    ## Check for arguments and set to defaults if missing ----------------------
+
+    tinify_opts <- .tinify_getset_opts(overwrite,
+                                       suffix,
+                                       quiet,
+                                       return_path,
+                                       resize)
+
+    ## Generic argument error checking -----------------------------------------
+
+    .tinify_error_check(overwrite = tinify_opts$overwrite,
+                        suffix = tinify_opts$suffix,
+                        quiet = tinify_opts$quiet,
+                        return_path = tinify_opts$return_path,
+                        resize = tinify_opts$resize)
+
+    # Main function body =======================================================
+
+    ## overwrite ---------------------------------------------------------------
+
+    if(!missing(overwrite)){
+      options(tinify.overwrite = overwrite)
+      cli::cli_alert_success("Tinify {.var overwrite} changed to: {.field {overwrite}}")
+    }
+
+    ## suffix ------------------------------------------------------------------
+
+    if(!missing(suffix)){
+      options(tinify.suffix = suffix)
+      cli::cli_alert_success("Tinify {.var suffix} changed to: {.field {suffix}}")
+    }
+
+    ## quiet -------------------------------------------------------------------
+
+    if(!missing(quiet)){
+      options(tinify.quiet = quiet)
+      cli::cli_alert_success("Tinify {.var quiet} changed to: {.field {quiet}}")
+    }
+
+    ## return_path -------------------------------------------------------------
+
+    if(!missing(return_path)){
+      options(tinify.return_path = return_path)
+      if(!is.null(return_path)) {
+        cli::cli_alert_success("Tinify {.var return_path} changed to: {.field \"{return_path}\"}")
+      } else {
+        cli::cli_alert_success("Tinify {.var return_path} changed to: {.field No return}")
+      }
+    }
+
+    ## resize ------------------------------------------------------------------
+
+    if(!missing(resize)){
+      options(tinify.resize = resize)
+      if(!is.null(resize)) {
+        for (i in 1:length(resize)) {
+            if (is.character(resize[[i]])) {
+              cli::cli_alert_success('Tinify {.var resize} {(names(resize)[i])} changed to: {.field \"{resize[[i]]}\"}')
+            } else {
+              cli::cli_alert_success('Tinify {.var resize} {(names(resize)[i])} changed to: {.field {resize[[i]]}}')
+            }
+          }
+      } else {
+        cli::cli_alert_success("Tinify {.var resize} changed to: {.field No resize}")
+      }
+    }
+
+  }
+
+}
+
+#' Prints the currently set tinify_defaults
+#' @noRd
+#' @keywords internal
+.tinify_print_defaults <- function() {
+
+    # overwrite ----------------------------------------------------------------
+
+    cli::cli_alert_info("Tinify {.var overwrite} default is: {.field {getOption('tinify.overwrite', default = FALSE)}}")
+
+    # suffix -------------------------------------------------------------------
+
+    cli::cli_alert_info("Tinify {.var suffix} default is: {.field \"{if (!is.null(getOption('tinify.suffix'))) getOption('tinify.suffix') else '_tiny'}\"}")
+
+    # quiet --------------------------------------------------------------------
+
+    cli::cli_alert_info("Tinify {.var quiet} default is: {.field {getOption('tinify.quiet', default = FALSE)}}")
+
+    # return_path --------------------------------------------------------------
+
+    cli::cli_alert_info("Tinify {.var return_path} default is: {.field {if (!is.null(getOption('tinify.return_path'))) getOption('tinify.return_path') else 'No return'}}")
+
+    # resize -------------------------------------------------------------------
+
     if(!is.null(getOption("tinify.resize"))) {
       resize <- getOption("tinify.resize")
       for (i in 1:length(resize)) {
-          if (is.character(resize[[i]])) {
-            message(glue::glue('Tinify \'resize\' {(names(resize)[i])} set to: \"{resize[[i]]}\"'))
-          } else {
-            message(glue::glue('Tinify \'resize\' {(names(resize)[i])} set to: {resize[[i]]}'))
-          }
+        if (is.character(resize[[i]])) {
+          cli::cli_alert_info('Tinify {.var resize} {(names(resize)[i])} default is: {.field \"{resize[[i]]}\"}')
+        } else {
+          cli::cli_alert_info('Tinify {.var resize} {(names(resize)[i])} default is: {.field {resize[[i]]}}')
         }
+      }
     } else {
-      message(glue::glue("Tinify 'resize' set to: No resize"))
+      cli::cli_alert_info("Tinify {.var resize} default is: {.field No resize}")
     }
-  }
 
 }